requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-api-python-client
google-auth-oauthlib
motor  # This is the MongoDB driver
pydantic[email]